+++ src/adminApp/customRoutes.js
@@ -17,7 +17,7 @@ import EncounterTypeEdit from "./EncounterType/EncounterTypeEdit";
 const CreateConcept = () => <CreateEditConcept isCreatePage={true} />;
 
 export default [
-  <Route exact path="/languages" component={CreateEditLanguages} />,
+  <Route exact path="/editLanguages" component={CreateEditLanguages} />,
   <Route exact path="/filters" component={CreateEditFilters} />,
   <Route exact path="/forms/:formUUID" component={FormDetails} />,
   <Route exact path="/concept/create" component={CreateConcept} />,
